requirements.txt
flake8==3.3.0
mockldap==0.2.8
pytest==3.0.7
pytest-django==3.1.2
pytest-pythonpath==0.7.1
